requirements.txt
aiohttp
gitpython
lxml
orjson
requests
tomli